removed useless AmazonEsSearch and gave a better name to the Repository
